Almost done with answer just add validations
