+++ app.js
@@ -9,6 +9,7 @@ const { sessionSecret } = require('./config');
 const SequelizeStore = require('connect-session-sequelize')(session.Store);
 const indexRouter = require('./routes/index');
 const usersRouter = require('./routes/users');
+const answersRouter = require('./routes/answers');
 const { restoreUser } = require('./auth');
 
 const app = express();
@@ -40,6 +41,7 @@ app.use(restoreUser);
 
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
+app.use('/answers', answersRouter);
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
