+++ views/answer-form.pug
@@ -2,15 +2,15 @@ extends layout.pug
 
 block content
   if errors
-    div(class='alert alert-danger' role='alert')
+    div(role='alert')
       p The following error(s) occurred:
       ul
         each error in errors
           li= error
-    div(class='alert alert-danger' role='alert')
+    div(role='alert')
       p The following error(s) occurred:
       pre= JSON.stringify(error, null, 2)
-  form(action='/users/answer/add' method='post')
+  form(action='/answers/add' method='post')
     input(type='hidden' name='_csrf' value=csrfToken)
     div(class='form-group')
       label(for='questionId') Question ID
