+++ views/answers.pug
@@ -20,7 +20,7 @@ block content
                         td= downvotes[answer.id]
                         td
                             //- a(href=`/users/answers/${answer.id}` role='button') Delete
-                            form(action=`/users/answers/${answer.id}` method='get')
+                            form(action=`/answers/${answer.id}` method='get')
                                 input(type='hidden' name='_csrf' value=csrfToken)
                                 button(type='submit') Delete
     else
